Added a selection field to define a default value for VAT on Payment flag.
Usually, invoices emitted for a particular customer/supplier have the
same value for the VAT on Payment flag.
This patch adds the possibility to specify the default value of such a
flag.
This definition must be done inside the partner form and is made through
a selection. This is done to force the user to flag the right value.
Using a boolean is too easy to forget to check the field.
Once the default value is set, all subsequent invoices created for that
partner have the VAT on Payment flag automatically set to the right value.