[ADD] Moved default_has_vat_on_payment on account.fiscal.position.
After some research, the VAT on Payment seems dependent on the account
fiscal position of the customer/supplier.
Moved the default_has_vat_on_payment flag to that model to take that
into account.
The field is now a boolean because is not needed anymore to force the
user to make a choice.