Generate a unique Slimpay order reference to avoid duplicate errors
Odoo tries to reuse the unused order numbers as much as possible, so do
not use the order name as a Slimpay reference: in case of an error on
Slimpay's side, the order reference would no more be available.

This makes it possible to completely remove error handling code in the
controller and a dedicated Slimpay helper.

This commit also tries to clarify the places where the chosen Slimpay
reference format is used, by factoring out build and parse functions.