[12.0][FIX] - account-payment: fix unit tests
There was a fix in `account_payment` module that changed the way how `multi` field is computed.
In the case of `test_2_two_invoices_register_payment` in `account_payment_multi_deduction` module,
it's evaluated to True and this made the payment registration `amount` field readonly.

https://github.com/odoo/odoo/blob/867de2bad3f6096098cb47611bcf0096b85e24b8/addons/account/models/account_payment.py#L303

Now, this operation raises an error.

https://github.com/OCA/account-payment/blob/4f7c56d8f501dde75a04f61076a27d2b011274c0/account_payment_multi_deduction/tests/test_payment_multi_deduction.py#L99

https://travis-ci.org/github/OCA/account-payment/jobs/658790207#L1020

multi=True caused this error too as the payment register is not grouping invoices

https://travis-ci.org/github/OCA/account-payment/jobs/658790207#L1109