[FIX] account_check_printing_report_base: Pin num2words version and fix typo in function call
The current version used and installed by Odoo is 0.5.6: https://github.com/odoo/odoo/blob/13.0/requirements.txt#L21
It has a typo in the function call to to_currency where "separator" is spelled "seperator".
As that is the version that Odoo uses, we should adapt to it, and also pin to ensure we are installing the correct one.

The typo is fixed in https://github.com/savoirfairelinux/num2words/commit/2eee037648eb2245d9714f694e4868b80082a216 for version 0.5.10, so we should revert this if we ever upgrade to that version

TT30932