Fix application of discount with cash rounding
If a cash rounding is applied, there can be two move lines for the
tax as a rounding one is added automatically.
The condition to store the financial discount must therefore check
if all the tax lines share the same tax. The financial discount
will then be linked with the bigger amount tax line in order to
generate the proper write off in the reconciliation.