[FIX] account_payment_promissory_note: fix flaky test
The onchange was not checking the `active_model` before browsing the `active_ids`.

At the end of the day, the upstream wizard already does all the checks and
fulfills the `line_ids` appropriately.

The flaky test was `test_3_due_date_propagation`, where the wizard was created
with active invoices, not active lines. Sometimes it worked when by chance the
id of lines somehow matched the id of the invoice. But it was flaky nevertheless.

@moduon MT-10470